--- src/pages/companies/CompanyTable.js
@@ -1,34 +1,0 @@
-import React from 'react';
-import Table from 'react-bootstrap/Table';
-
-export default function CompanyTable() {
-  return (
-    <div>
-      <Table striped bordered hover>
-      <thead>
-        <tr>
-          <th>ID</th>
-          <th>FullName</th>
-          <th>CompanyName</th>
-          <th>Status</th>
-          <th></th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr>
-          <td></td>
-          <td></td>
-          <td></td>
-          <td></td>  
-        </tr>
-        <tr>
-          <td></td>
-          <td></td>
-          <td></td>
-          <td></td>  
-        </tr>
-      </tbody>
-    </Table>
-    </div>
-  )
-}
