+++ src/pages/home/Home.css
@@ -40,7 +40,8 @@
  
   .alert{
     width:70%;
-    top: 100px;
+    top: 200px;
+    
   }
 
   .Home{
