--- src/pages/people/TablePeople.js
@@ -1,29 +1,0 @@
-import React from 'react'
-import Table from 'react-bootstrap/Table';
-
-export default function() {
-  return (
-    <div>
-    <Table striped bordered hover>
-      <thead>
-        <tr>
-          <th>FullName</th>
-          <th>CompanyName</th>
-          <th>Position</th>
-          
-        </tr>
-      </thead>
-      <tbody>
-        <tr>
-          <td></td>
-          <td></td>
-          <td></td>
-          
-        </tr>
-       
-      </tbody>
-    </Table>
-
-   </div>
-  )
-}
