--- src/pages/project/CreateProject.js
@@ -1,32 +1,0 @@
-import React, { useState } from "react";
-import Button from "react-bootstrap/Button";
-import Modal from "react-bootstrap/Modal";
-import Create from "../create/Create";
-
-function Example1() {
-  const [lgShow, setLgShow] = useState(false);
-
-  return (
-    <>
-      <Button onClick={() => setLgShow(true)}>+Create New</Button>
-
-      <Modal
-        size="lg"
-        show={lgShow}
-        onHide={() => setLgShow(false)}
-        aria-labelledby="example-modal-sizes-title-lg"
-      >
-        <Modal.Header closeButton>
-          <Modal.Title id="example-modal-sizes-title-lg">
-            Create Project
-          </Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          <Create />
-        </Modal.Body>
-      </Modal>
-    </>
-  );
-}
-
-export default Example1;
