--- src/pages/project/ProjectTable.js
@@ -1,33 +1,0 @@
-import React from 'react'
-import Table from 'react-bootstrap/Table';
-
-export default function ProjectTable() {
-  return (
-    <div>
-      <Table striped bordered hover>
-      <thead>
-        <tr>
-          <th>ID</th>
-          <th>Name</th>
-          <th>Status</th>
-          <th>Lead</th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr>
-          <td></td>
-          <td></td>
-          <td></td>
-          <td></td>  
-        </tr>
-        <tr>
-          <td></td>
-          <td></td>
-          <td></td>
-          <td></td>  
-        </tr>
-      </tbody>
-    </Table>
-    </div>
-  )
-}
